+++ xarray/coding/variables.py
@@ -1,4 +1,5 @@
 """Coders for individual Variable objects."""
+from typing import Any
 import warnings
 from functools import partial
 
@@ -124,11 +125,12 @@ def pop_to(source, dest, key, name=None):
     return value
 
 
-def _apply_mask(data,  # type: np.ndarray
-                encoded_fill_values,  # type: list
-                decoded_fill_value,  # type: Any
-                dtype,  # type: Any
-                ):  # type: np.ndarray
+def _apply_mask(
+    data: np.ndarray,
+    encoded_fill_values: list,
+    decoded_fill_value: Any,
+    dtype: Any,
+) -> np.ndarray:
     """Mask all matching values in a NumPy arrays."""
     data = np.asarray(data, dtype=dtype)
     condition = False