+++ xarray/tests/test_dataset.py
@@ -4,6 +4,7 @@ import warnings
 from collections import OrderedDict
 from copy import copy, deepcopy
 from io import StringIO
+import pickle
 from textwrap import dedent
 
 import numpy as np
@@ -24,10 +25,6 @@ from . import (
     raises_regex, requires_bottleneck, requires_dask, requires_scipy,
     source_ndarray)
 
-try:
-    import cPickle as pickle
-except ImportError:
-    import pickle
 try:
     import dask.array as da
 except ImportError:
@@ -3043,11 +3040,25 @@ class TestDataset(object):
         # check roundtrip
         assert_identical(ds, Dataset.from_dict(actual))
 
-        # verify coords are included roundtrip
-        expected = ds.set_coords('b')
-        actual = Dataset.from_dict(expected.to_dict())
+        # check the data=False option
+        expected_no_data = expected.copy()
+        del expected_no_data['coords']['t']['data']
+        del expected_no_data['data_vars']['a']['data']
+        del expected_no_data['data_vars']['b']['data']
+        expected_no_data['coords']['t'].update({'dtype': '<U1',
+                                                'shape': (10,)})
+        expected_no_data['data_vars']['a'].update({'dtype': 'float64',
+                                                   'shape': (10,)})
+        expected_no_data['data_vars']['b'].update({'dtype': 'float64',
+                                                   'shape': (10,)})
+        actual_no_data = ds.to_dict(data=False)
+        assert expected_no_data == actual_no_data
 
-        assert_identical(expected, actual)
+        # verify coords are included roundtrip
+        expected_ds = ds.set_coords('b')
+        actual = Dataset.from_dict(expected_ds.to_dict())
+
+        assert_identical(expected_ds, actual)
 
         # test some incomplete dicts:
         # this one has no attrs field, the dims are strings, and x, y are
@@ -4428,6 +4439,48 @@ def ds(request):
                         'time': ('time', np.linspace(0, 1.0, 10)),
                         'c': ('y', ['a', 'b']),
                         'y': range(2)})
+
+
+@pytest.mark.parametrize('dask', [True, False])
+@pytest.mark.parametrize(('boundary', 'side'), [
+    ('trim', 'left'), ('pad', 'right')])
+def test_coarsen(ds, dask, boundary, side):
+    if dask and has_dask:
+        ds = ds.chunk({'x': 4})
+
+    actual = ds.coarsen(time=2, x=3, boundary=boundary, side=side).max()
+    assert_equal(
+        actual['z1'],
+        ds['z1'].coarsen(time=2, x=3, boundary=boundary, side=side).max())
+    # coordinate should be mean by default
+    assert_equal(actual['time'], ds['time'].coarsen(
+        time=2, x=3, boundary=boundary, side=side).mean())
+
+
+@pytest.mark.parametrize('dask', [True, False])
+def test_coarsen_coords(ds, dask):
+    if dask and has_dask:
+        ds = ds.chunk({'x': 4})
+
+    # check if coord_func works
+    actual = ds.coarsen(time=2, x=3, boundary='trim',
+                        coord_func={'time': 'max'}).max()
+    assert_equal(actual['z1'],
+                 ds['z1'].coarsen(time=2, x=3, boundary='trim').max())
+    assert_equal(actual['time'],
+                 ds['time'].coarsen(time=2, x=3, boundary='trim').max())
+
+    # raise if exact
+    with pytest.raises(ValueError):
+        ds.coarsen(x=3).mean()
+    # should be no error
+    ds.isel(x=slice(0, 3 * (len(ds['x']) // 3))).coarsen(x=3).mean()
+
+    # working test with pd.time
+    da = xr.DataArray(
+        np.linspace(0, 365, num=364), dims='time',
+        coords={'time': pd.date_range('15/12/1999', periods=364)})
+    actual = da.coarsen(time=2).mean()
 
 
 def test_rolling_properties(ds):